+++ src/legacy/ModalDialog/index.ts
@@ -1,0 +1,1 @@
+export { ModalDialog as LegacyModalDialog } from './ModalDialog';